+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Fixed bug on negative partition id not being rejected (get partition, send to partition and get offsets operations), by adding a `minimum` constraint to the `partitionid` parameter in the OpenAPI spec.
 * Separated HTTP server and SSL/TLS configuration to use the Vert.x `ServerSSLOptions` API, enabling future support for configuring TLS key exchange groups (e.g., for post-quantum cryptography).
+* Added `http.ssl.named.groups` configuration parameter to specify the SSL/TLS named groups to be used on the HTTP server side of the bridge. If not set, the JVM defaults are used.
 
 ## 1.1.0